Fix #10, fix #11, fix #13 & fix #14

+++ installer/deploy.sh
@@ -1,45 +1,40 @@
 #!/bin/bash
 
-INSTALL_PATH='/home/installer'
+REPO_PATH='/home/vmail'
 
-## Config file
-source ./config.conf
+if [ -f $REPO_PATH/installer/config.conf ]; then
 
-
-## www
-mkdir -p /var/www
-chown -R http:http /var/www/
-sed -i -e "s/CONFIG_MARIADB_WWW_PASSWORD/$CONFIG_MARIADB_WWW_PASSWORD/g" /var/www/app/config/parameters.yml
-cd /var/www
-
-# Cache and logs
-rm -r app/cache/*
-rm -r app/logs/*
-# Managed by Composer
-rm -r app/bootstrap.php.cache
-rm -r bin/
-rm -r vendor/
-# Assets
-rm -r web/bundles/
-
-php composer.phar self-update
-php composer.phar update
-php app/console assets:install
-php composer.phar dump-autoload --optimize
-php app/console doctrine:schema:update --force
-chmod -R 777 app/cache
-chmod -R 777 app/logs
+    ## Config file
+    chmod 600 $REPO_PATH/installer/config.conf
+    source $REPO_PATH/installer/config.conf
 
 
-## Console scripts
-cp $INSTALL_PATH/console/* /var/www
-chmod 777 /var/www/*.sh
-sed -i -e "s/CONFIG_IP_PRIMARY/$CONFIG_IP_PRIMARY/g" /var/www/dfilter.sh
+    ## www
+    cp -r $REPO_PATH/www /var/www
+    sed -i -e "s/CONFIG_MARIADB_WWW_PASSWORD/$CONFIG_MARIADB_WWW_PASSWORD/g" /var/www/app/config/parameters.yml.dist
+    cd /var/www
+    php composer.phar self-update
+    php composer.phar update
+    php app/console assets:install
+    php composer.phar dump-autoload --optimize
+    php app/console doctrine:schema:update --force
 
 
-## Symbolic links
-ln -s /usr/share/webapps/roundcubemail/ /var/www/web/webmail
-ln -s /usr/share/webapps/piwik/ /var/www/web/piwik
+    ## Virtual mailbox domain
+    mysql -u root -p"$CONFIG_MARIADB_ROOT_PASSWORD" -e "INSERT INTO vmailme.domain SET name=\"$CONFIG_DOMAIN\";"
 
 
-echo "Deploy www [OK]."
+    ## Console scripts
+    cp $REPO_PATH/installer/console/* /var/www
+    chmod 777 /var/www/*.sh
+    sed -i -e "s/CONFIG_IP_PRIMARY/$CONFIG_IP_PRIMARY/g" /var/www/dfilter.sh
+
+
+    ## Symbolic links
+    ln -s /usr/share/webapps/roundcubemail/ /var/www/web/webmail
+    ln -s /usr/share/webapps/piwik/ /var/www/web/piwik
+
+
+    chown -R http:http /var/www/
+    echo "Deploy www [OK]."
+fi
